+++ src/components/statusLabel/StatusLabel.tsx
@@ -9,8 +9,8 @@ import theme from './theme.css';
 export interface StatusLabelProps extends Omit<BoxProps, 'size'> {
   children: ReactNode;
   className?: string;
-  color?: Exclude<typeof COLORS[number], 'teal'>;
-  size?: Exclude<typeof SIZES[number], 'tiny' | 'large' | 'fullscreen' | 'smallest' | 'hero'>;
+  color?: Exclude<(typeof COLORS)[number], 'teal'>;
+  size?: Exclude<(typeof SIZES)[number], 'tiny' | 'large' | 'fullscreen' | 'smallest' | 'hero'>;
 }
 
 const StatusLabel: GenericComponent<StatusLabelProps> = forwardRef<HTMLElement, StatusLabelProps>(